% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cond_case.R
\name{case}
\alias{case}
\title{A powerful replacement for base::switch}
\usage{
case(obj, ...)
}
\arguments{
\item{obj}{An object/value.}

\item{...}{An odd number of expressions. A set of test/expression
pairs plus an expression to be evalauted if all tests fail. Equivalent
of nested \code{if (identical(obj, a)) b else c} calls.}
}
\description{
See \url{https://clojuredocs.org/clojure.core/case}.
}
\examples{
case(x,
     list(1,2), "abc",
     c('a','b'), 123,
     NULL, "oh no!",
     "something else")
# Is transformed to:
# if (identical(x, list(1, 2))) "abc" else
#     if (identical(x, c("a","b"))) 123 else
#         if (identical(x, NULL)) "oh no!" else
#             "something else"
}

