% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cond_case.R
\name{cond}
\alias{cond}
\title{Equivalent of nested if ... else ... calls}
\usage{
cond(...)
}
\arguments{
\item{...}{An odd number of expressions. A set of test/expression
pairs plus an expression to be evalauted if all tests fail. Equivalent
of nested \code{if ... else ...} calls.}
}
\description{
See \url{https://clojuredocs.org/clojure.core/cond}.
}
\examples{
cond(x < 5, 'a',
     x < 2, 'b',
     x > 1, 'c',
            'd')
# Is transformed to:
# if (x < 5) "a" else
#     if (x < 2) "b" else
#         if (x > 1) "c" else "d"
}

