% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cond_thread.R
\name{condv->}
\alias{condv->}
\title{A vectorised version of clojR::`cond->`}
\usage{
"condv->"(obj, ...)
}
\arguments{
\item{obj}{An object/value.}

\item{...}{An even number of expressions. A set of test/expression pairs.}
}
\description{
See \code{\link[clojR]{`cond->`}}. This function (\code{`condv->`})
uses internally \code{ifelse} instead of \code{if}.
}

