% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cond_case.R
\name{condv}
\alias{condv}
\title{Equivalent of nested ifelse calls (vectorised clojR::cond)}
\usage{
condv(...)
}
\arguments{
\item{...}{An odd number of expressions. A set of test/expression
pairs plus an expression to be evalauted if all tests fail. Equivalent
of nested \code{\link[base]{ifelse}} calls.}
}
\description{
See \url{https://clojuredocs.org/clojure.core/cond}.
}
\examples{
condv(x < 5, 'a',
      x < 2, 'b',
      x > 1, 'c',
             'd')
# Is transformed to:
# ifelse(x < 5, "a",
#        ifelse(x < 2, "b",
#               ifelse(x > 1, "c",
#                      "d")))
}

