% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doc.R
\name{doc}
\alias{doc}
\title{Display in-body function documentation (a "docstring")}
\usage{
doc(f)
}
\arguments{
\item{f}{A function.}
}
\description{
See \url{https://clojuredocs.org/clojure.repl/doc}.
The "docstring" is extracted from the comments starting
with a double hash character (##).
}
\examples{
fff <- function(x = 1) {
    ## This is my function:
    ## argument x, default 1.
    ## returns x + 10.
    x + 10
}
doc(fff)
# displays in the console:

# This is my function:
# argument x, default 1.
# returns x + 10
}

