% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thread.R
\name{as->}
\alias{as->}
\title{Threading a value, bound to a symbol, through expressions}
\usage{
"as->"(value, symbol, ...)
}
\arguments{
\item{value}{An initial value.}

\item{symbol}{An unquoted symbol name which will be replaced
by the values returned by the previous expressions.}

\item{...}{The expressions, most likely including the
symbol defined in \code{symbol} (possibly nested).}
}
\description{
See \url{https://clojuredocs.org/clojure.core/as->}.
}
\examples{
`as->`(1000, nn,
       mean(c(1,nn)),
       sum(c(nn,5)),
       `-`(nn,13))
# Is transformed to:
# sum(c(mean(c(1, 1000)), 5)) - 13
# an evaluates to 492.5.
}

