% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/thread.R
\name{->}
\alias{->}
\title{Threading a value, as the first argument, through expressions}
\usage{
"->"(value, ...)
}
\arguments{
\item{value}{An initial value.}

\item{...}{The expressions.}
}
\description{
See \url{https://clojuredocs.org/clojure.core/->}.
}
\examples{
`->`(3,
     `-`(400),
     sum(7,8))
# Is transformed to:
# sum(3 - 400, 7, 8)
# an evaluates to -382.
}

